+++ react/features/base/lib-jitsi-meet/native/index.js
@@ -1,2 +1,1 @@
-import './polyfills-browser';
 import './WiFiStats';
