--- react/features/base/storage/_.native.js
@@ -1,1 +1,0 @@
-export * from './native';
