+++ react/features/base/storage/index.js
@@ -1,4 +1,3 @@
-export * from './_';
 export { default as PersistenceRegistry } from './PersistenceRegistry';
 
 import './middleware';
