--- react/features/base/storage/native/index.js
@@ -1,1 +1,0 @@
-import './polyfills-browser';
