+++ react/features/mobile/polyfills/Storage.js
