+++ react/features/mobile/polyfills/browser.js
@@ -1,7 +1,9 @@
+import { Platform } from 'react-native';
 import BackgroundTimer from 'react-native-background-timer';
+
 import '@webcomponents/url'; // Polyfill for URL constructor
 
-import { Platform } from '../../react';
+import Storage from './Storage';
 
 /**
  * Gets the first common prototype of two specified Objects (treating the
@@ -372,7 +374,7 @@ function _visitNode(node, callback) {
     }
 
     // WebRTC
-    require('./polyfills-webrtc');
+    require('./webrtc');
 
     // CallStats
     //
@@ -417,5 +419,19 @@ function _visitNode(node, callback) {
     global.clearInterval = BackgroundTimer.clearInterval.bind(BackgroundTimer);
     global.setInterval = BackgroundTimer.setInterval.bind(BackgroundTimer);
     global.setTimeout = (fn, ms = 0) => BackgroundTimer.setTimeout(fn, ms);
+
+    // localStorage
+    if (typeof global.localStorage === 'undefined') {
+        global.localStorage = new Storage('@jitsi-meet/');
+    }
+
+    // sessionStorage
+    //
+    // Required by:
+    // - herment
+    // - Strophe
+    if (typeof global.sessionStorage === 'undefined') {
+        global.sessionStorage = new Storage();
+    }
 
 })(global || window || this); // eslint-disable-line no-invalid-this
