+++ react/features/mobile/polyfills/bundler.js
