+++ react/features/mobile/polyfills/index.js
@@ -1,0 +1,2 @@
+import './bundler';
+import './browser';
