+++ react/features/mobile/polyfills/webrtc.js
